--- src/main/java/cloneproject/Instagram/dto/MessageResponse.java
@@ -1,15 +1,0 @@
-package cloneproject.Instagram.dto;
-
-
-import lombok.Getter;
-
-@Getter
-public class MessageResponse {
-    
-    private String message;
-
-    public MessageResponse(String message){
-        this.message = message;
-    }
-
-}
