+++ src/main/java/cloneproject/Instagram/dto/member/ReissueRequest.java
@@ -1,4 +1,4 @@
-package cloneproject.Instagram.dto;
+package cloneproject.Instagram.dto.member;
 
 import javax.validation.constraints.NotBlank;
